provided more info on errors

+++ src/Dodkin/MachineManagementInfo.cs
@@ -14,13 +14,13 @@
             this.properties = properties;
         }
 
-        public string[] ActiveQueueFormatNames => this.properties.GetValue<string[]>(MQ.PROPID.MGMT_MSMQ.ACTIVEQUEUES);
-        public string[] PrivateQueuePathNames => this.properties.GetValue<string[]>(MQ.PROPID.MGMT_MSMQ.PRIVATEQ);
-        public string DirectoryService => this.properties.GetValue<string>(MQ.PROPID.MGMT_MSMQ.DSSERVER);
-        public bool IsConnected => String.Equals(this.properties.GetValue<string>(MQ.PROPID.MGMT_MSMQ.CONNECTED),
+        public string[] ActiveQueueFormatNames => this.properties.GetValue<string[]>((int)MQ.PROPID.MGMT_MSMQ.ACTIVEQUEUES);
+        public string[] PrivateQueuePathNames => this.properties.GetValue<string[]>((int)MQ.PROPID.MGMT_MSMQ.PRIVATEQ);
+        public string DirectoryService => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_MSMQ.DSSERVER);
+        public bool IsConnected => String.Equals(this.properties.GetValue<string>((int)MQ.PROPID.MGMT_MSMQ.CONNECTED),
                 MSMQ_CONNECTED, StringComparison.OrdinalIgnoreCase);
-        public string Type => this.properties.GetValue<string>(MQ.PROPID.MGMT_MSMQ.TYPE);
-        public long QueueSize => (long)this.properties.GetValue<ulong>(MQ.PROPID.MGMT_MSMQ.BYTES_IN_ALL_QUEUES);
+        public string Type => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_MSMQ.TYPE);
+        public long QueueSize => (long)this.properties.GetValue<ulong>((int)MQ.PROPID.MGMT_MSMQ.BYTES_IN_ALL_QUEUES);
 
         public void Dispose()
         {
