+++ src/Dodkin/QueueManagementInfo.cs
@@ -15,21 +15,21 @@
             this.properties = properties;
         }
 
-        public string PathName => this.properties.GetValue<string>(MQ.PROPID.MGMT_QUEUE.PATHNAME);
-        public string FormatName => this.properties.GetValue<string>(MQ.PROPID.MGMT_QUEUE.FORMATNAME);
-        public string Type => this.properties.GetValue<string>(MQ.PROPID.MGMT_QUEUE.TYPE);
-        public string Location => this.properties.GetValue<string>(MQ.PROPID.MGMT_QUEUE.LOCATION);
-        public bool? IsTransactional => GetBoolean(MQ.PROPID.MGMT_QUEUE.XACT);
-        public bool? IsForeign => GetBoolean(MQ.PROPID.MGMT_QUEUE.FOREIGN);
-        public int MessageCount => (int)this.properties.GetValue<uint>(MQ.PROPID.MGMT_QUEUE.MESSAGE_COUNT);
-        public int Size => (int)this.properties.GetValue<uint>(MQ.PROPID.MGMT_QUEUE.BYTES_IN_QUEUE);
-        public int JournalMessageCount => (int)this.properties.GetValue<uint>(MQ.PROPID.MGMT_QUEUE.JOURNAL_MESSAGE_COUNT);
-        public int JournalSize => (int)this.properties.GetValue<uint>(MQ.PROPID.MGMT_QUEUE.BYTES_IN_JOURNAL);
-        public string State => this.properties.GetValue<string>(MQ.PROPID.MGMT_QUEUE.STATE);
-        public string[] NextHops => this.properties.GetValue<string[]>(MQ.PROPID.MGMT_QUEUE.NEXTHOPS);
+        public string PathName => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_QUEUE.PATHNAME);
+        public string FormatName => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_QUEUE.FORMATNAME);
+        public string Type => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_QUEUE.TYPE);
+        public string Location => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_QUEUE.LOCATION);
+        public bool? IsTransactional => GetBoolean((int)MQ.PROPID.MGMT_QUEUE.XACT);
+        public bool? IsForeign => GetBoolean((int)MQ.PROPID.MGMT_QUEUE.FOREIGN);
+        public int MessageCount => (int)this.properties.GetValue<uint>((int)MQ.PROPID.MGMT_QUEUE.MESSAGE_COUNT);
+        public int Size => (int)this.properties.GetValue<uint>((int)MQ.PROPID.MGMT_QUEUE.BYTES_IN_QUEUE);
+        public int JournalMessageCount => (int)this.properties.GetValue<uint>((int)MQ.PROPID.MGMT_QUEUE.JOURNAL_MESSAGE_COUNT);
+        public int JournalSize => (int)this.properties.GetValue<uint>((int)MQ.PROPID.MGMT_QUEUE.BYTES_IN_JOURNAL);
+        public string State => this.properties.GetValue<string>((int)MQ.PROPID.MGMT_QUEUE.STATE);
+        public string[] NextHops => this.properties.GetValue<string[]>((int)MQ.PROPID.MGMT_QUEUE.NEXTHOPS);
 
-        public int SubqueueCount => (int)this.properties.GetValue<uint>(MQ.PROPID.MGMT_QUEUE.SUBQUEUE_COUNT);
-        public string[] SubqueueNames => this.properties.GetValue<string[]>(MQ.PROPID.MGMT_QUEUE.SUBQUEUE_NAMES);
+        public int SubqueueCount => (int)this.properties.GetValue<uint>((int)MQ.PROPID.MGMT_QUEUE.SUBQUEUE_COUNT);
+        public string[] SubqueueNames => this.properties.GetValue<string[]>((int)MQ.PROPID.MGMT_QUEUE.SUBQUEUE_NAMES);
 
         public void Dispose()
         {
